+++ src/main/java/com/foolsix/fancyenchantments/enchantment/BubbleShield.java
@@ -1,0 +1,30 @@
+package com.foolsix.fancyenchantments.enchantment;
+
+import com.foolsix.fancyenchantments.FancyEnchantments;
+import com.foolsix.fancyenchantments.enchantment.EssentiaEnch.AquaEnchantment;
+import com.foolsix.fancyenchantments.util.ModConfig;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import net.minecraftforge.event.entity.living.LivingHurtEvent;
+
+public class BubbleShield extends AquaEnchantment {
+    public static final String NAME = "bubble_shield";
+    private static final ModConfig.BubbleShieldOptions CONFIG = FancyEnchantments.getConfig().bubbleShieldOptions;
+
+    public BubbleShield() {
+        super(CONFIG, EnchantmentCategory.ARMOR_CHEST, new EquipmentSlot[]{EquipmentSlot.CHEST});
+    }
+
+    public void reduceDamage(LivingHurtEvent e) {
+        if (e.getEntity() instanceof Player player) {
+            int level = EnchantmentHelper.getEnchantmentLevel(this, player);
+            if (level > 0 && player.getAirSupply() >= player.getMaxAirSupply()) {
+                int reducer = (int) (player.getAirSupply() * CONFIG.costRatio * CONFIG.multiplier * level);
+                player.setAirSupply((int) Math.max(0, player.getAirSupply() * (1 - CONFIG.costRatio)));
+                e.setAmount(Math.max(0f, e.getAmount() - reducer));
+            }
+        }
+    }
+}