+++ src/main/java/com/foolsix/fancyenchantments/enchantment/util/EnchantmentReg.java
@@ -97,16 +97,21 @@ public class EnchantmentReg {
 
     public static final RegistryObject<Enchantment> SELF_IMMOLATION = reg(SelfImmolation.NAME, SelfImmolation::new);
 
-    public static void register(IEventBus eventBus){
+    public static final RegistryObject<Enchantment> BUBBLE_SHIELD = reg(BubbleShield.NAME, BubbleShield::new);
+
+    public static final RegistryObject<Enchantment> DROWNING = reg(Drowning.NAME, Drowning::new);
+
+    public static final RegistryObject<Enchantment> PERVERT = reg(Pervert.NAME, Pervert::new);
+
+    public static void register(IEventBus eventBus) {
         ENCHANTMENTS.register(eventBus);
     }
 
     public static RegistryObject<Enchantment> reg(String name, Supplier<? extends Enchantment> sup) {
-        String lang = String.format("\"enchantment.fancyenchantments.%s\": \"%s\",", name, EnchUtils.getLangName(name));
-        String desc = String.format("\"enchantment.fancyenchantments.%s.desc\": \"\",", name);
-        System.out.println(lang);
-        System.out.println(desc);
-
+//        String lang = String.format("\"enchantment.fancyenchantments.%s\": \"%s\",", name, EnchUtils.getLangName(name));
+//        String desc = String.format("\"enchantment.fancyenchantments.%s.desc\": \"\",", name);
+//        System.out.println(lang);
+//        System.out.println(desc);
         return ENCHANTMENTS.register(name, sup);
     }
 }