+++ src/main/java/com/foolsix/fancyenchantments/util/ModConfig.java
@@ -104,6 +104,16 @@ public class ModConfig implements ConfigData {
     public final LavaBurstOptions lavaBurstOptions = new LavaBurstOptions();
     @ConfigEntry.Gui.CollapsibleObject
     public final ArmorForgingOptions armorForgingOptions = new ArmorForgingOptions();
+    @ConfigEntry.Gui.CollapsibleObject
+    public final NirvanaOptions nirvanaOptions = new NirvanaOptions();
+    @ConfigEntry.Gui.CollapsibleObject
+    public final SelfImmolationOptions selfImmolationOptions = new SelfImmolationOptions();
+    @ConfigEntry.Gui.CollapsibleObject
+    public final BubbleShieldOptions bubbleShieldOptions = new BubbleShieldOptions();
+    @ConfigEntry.Gui.CollapsibleObject
+    public final DrowningOptions drowningOptions = new DrowningOptions();
+    @ConfigEntry.Gui.CollapsibleObject
+    public final PervertOptions pervertOptions = new PervertOptions();
 
 
     public static class ElementStatOptions {
@@ -116,7 +126,7 @@ public class ModConfig implements ConfigData {
         @Comment("actually twisted level - holy level")
         public int twistedLevelToGetDebuff = 10;
         @Comment("The probability of getting a debuff per second")
-        public float probabilityToGetDebuff = 0.02f;
+        public double probabilityToGetDebuff = 0.02;
         @Comment("second")
         public int debuffDuration = 3;
     }
@@ -579,6 +589,56 @@ public class ModConfig implements ConfigData {
             super(3, Rarity.VERY_RARE, 0.1);
             elementalCondition[Element.IGNIS.ordinal()] = 8;
             elementalCondition[Element.TERRA.ordinal()] = 8;
+        }
+    }
+
+    public static class NirvanaOptions extends BaseOptions {
+        @Comment("Minimum health ratio to trigger the effect")
+        public double minimumHealthRatio = 0.5;
+
+        NirvanaOptions() {
+            super(3, Rarity.RARE);
+        }
+    }
+
+    public static class SelfImmolationOptions extends BaseOptions {
+        public double baseProbability = 0.05;
+        @Comment("Probability of being ignited if target is on fire")
+        public double probabilityWithFire = 0.5;
+        @Comment("Second")
+        public int fireTime = 3;
+
+        SelfImmolationOptions() {
+            super(1, Rarity.RARE);
+        }
+    }
+
+    public static class BubbleShieldOptions extends BaseOptions {
+        @Comment("Triggering condition: airSupplyValue >= maxValue * Ratio")
+        public double airSupplyRatio = 0.8;
+        public double costRatio = 0.7;
+        @Comment("damage -= airSupplyValue * costRatio * multiplier * level")
+        public double multiplier = 0.1;
+
+        BubbleShieldOptions() {
+            super(3, Rarity.RARE);
+        }
+    }
+
+    public static class DrowningOptions extends BaseOptions {
+        public int airSupplyValue = -40;
+        public float damage = 1.0f;
+
+        DrowningOptions() {
+            super(1, Rarity.RARE);
+        }
+    }
+
+    public static class PervertOptions extends BaseOptions {
+        public double dropProbability = 0.9;
+
+        PervertOptions() {
+            super(1, Rarity.VERY_RARE);
         }
     }
 }